+++ CHANGELOG.md
@@ -11,6 +11,8 @@ New functionality:
 Improvements:
 * [OLMIS-3773](https://openlmis.atlassian.net/browse/OLMIS-3773): Adjusted facility search by ids and findAll after changing return type to page.
 * [OLMIS-5000](https://openlmis.atlassian.net/browse/OLMIS-5000): Refactored /requisitionsForConvert endpoint to improve its performance.
+* [OLMIS-5909](https://openlmis.atlassian.net/browse/OLMIS-5909): Adjusted permission checks in the service
+  *  a user without correct permission strings but with correct role assignments will be able to go through requisition workflow and find correct requisitions by the search endpoint.
 
 Bug fixes:
 * [OLMIS-5847](https://openlmis.atlassian.net/browse/OLMIS-5847): Made that requisition template names are case insensitive unique.